separate interfaces into their own folder
