+++ internal/interfaces/entity_deleter.go
@@ -1,0 +1,5 @@
+package interfaces
+
+type EntityDeleter interface {
+	DeleteEntity(id int, expectedTypes []string) error
+}
