+++ internal/interfaces/entity_getter.go
@@ -1,0 +1,7 @@
+package interfaces
+
+import "dns-api-go/internal/models"
+
+type EntityGetter interface {
+	GetEntity(id int, includeHA bool) (*models.Entity, error)
+}
